+++ CHANGELOG.md
@@ -7,5 +7,6 @@
 - readme: use astronvim4 as the config location
 - dev: gitignore inclusive patterns
 - lazy: include zipPlugin to support Clojure LSP jar analysis
+- plugin: astrocore enable spell & wrap options, set guifont for neovide
 
 ## Changed